pyproject.toml
[project]
name = "airzone-mqtt-hass"
version = "0.0.1"
description = "Airzone MQTT to Home Assistant"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
keywords = ["airzone", "hvac", "home", "home-assistant", "mqtt"] 
authors = [
  {name = "Álvaro Fernández Rojas", email = "noltari@gmail.com" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Home Automation",
]
dependencies = [
  "paho-mqtt",
]

[project.scripts]
airzone-mqtt-hass = "airzone_mqtt_hass:main"

[project.urls]
"Homepage" = "https://github.com/Noltari/airzone-mqtt-hass"
"Bug Tracker" = "https://github.com/Noltari/airzone-mqtt-hass/issues"

[tool.mypy]
python_version = "3.12"

[tool.pylint.MAIN]
py-version = "3.12"

[tool.pylint.BASIC]
class-const-naming-style = "any"

[tool.pylint."MESSAGES CONTROL"]
# Reasons disabled:
# duplicate-code - unavoidable
# invalid-name - not using snake case naming style
# too-many-* - are not enforced for the sake of readability
# too-few-* - same as too-many-*
disable = [
    "duplicate-code",
    "invalid-name",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-lines",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-positional-arguments",
    "too-many-public-methods",
    "too-many-statements",
]

[tool.pylint.REPORTS]
score = false

[tool.pylint.FORMAT]
expected-line-ending-format = "LF"

[tool.ruff.lint]
select = [
    "D",  # docstrings
    "E",  # pycodestyle
    "G",  # flake8-logging-format
    "I",  # isort
    "W",  # pycodestyle
]

ignore = [
    "D202",  # No blank lines allowed after function docstring
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
    "D401",  # First line should be in imperative mood
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
combine-as-imports = true
split-on-trailing-comma = false

[tool.setuptools]
platforms = ["any"]
zip-safe = false
include-package-data = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
